(PC-13826)[ADAGE] refactor: move function inside utils folder
