+++ adage-front/src/app/components/OffersInstantSearch/OffersSearch/Offers/Offer.tsx
@@ -4,36 +4,13 @@ import React, { useState } from 'react'
 
 import { ReactComponent as ChevronIcon } from 'assets/chevron.svg'
 import { ReactComponent as Logo } from 'assets/logo-without-text.svg'
-import { OfferType, VenueType } from 'utils/types'
+import { OfferType } from 'utils/types'
 
 import ContactButton from './ContactButton'
 import OfferDetails from './OfferDetails/OfferDetails'
 import OfferSummary from './OfferSummary/OfferSummary'
 import PrebookingButton from './PrebookingButton/PrebookingButton'
-
-const formatToReadableString = (input: string | null): string | null => {
-  if (input == null) {
-    return input
-  }
-  const lowerCasedInput = input.toLowerCase()
-  return lowerCasedInput.charAt(0).toUpperCase() + lowerCasedInput.slice(1)
-}
-
-const getOfferVenueAndOffererName = (offerVenue: VenueType): string => {
-  const venueName =
-    offerVenue.publicName || formatToReadableString(offerVenue.name)
-  const offererName = offerVenue.managingOfferer.name
-
-  if (venueName?.toLowerCase() === offererName.toLowerCase()) {
-    return venueName
-  }
-
-  if (!venueName) {
-    return offererName
-  }
-
-  return `${venueName} - ${offererName}`
-}
+import { getOfferVenueAndOffererName } from './utils/getOfferVenueAndOffererName'
 
 export const Offer = ({
   offer,
