+++ adage-front/src/app/components/OffersInstantSearch/OffersSearch/Offers/utils/getOfferVenueAndOffererName.ts
@@ -1,0 +1,19 @@
+import { VenueType } from 'utils/types'
+
+const formatToReadableString = (input: string): string => {
+  const lowerCasedInput = input.toLowerCase()
+  return lowerCasedInput.charAt(0).toUpperCase() + lowerCasedInput.slice(1)
+}
+
+export const getOfferVenueAndOffererName = (offerVenue: VenueType): string => {
+  const venueName =
+    offerVenue.publicName || formatToReadableString(offerVenue.name)
+
+  const offererName = offerVenue.managingOfferer.name
+
+  if (venueName.toLowerCase() === offererName.toLowerCase()) {
+    return venueName
+  }
+
+  return `${venueName} - ${offererName}`
+}
